Update patch set 1
Patch Set 1:

It's not my decision as I'm not a dev on this device, but  this flag affects everything that is built from this point forward:
- updater-binaries
- recovery
- and the actual ROM

Anything that includes the static code for libext4_utils_static or links to libext4_utils.so

If you build a recovery using the repo with this flag on, it will not perform a secure erase.  And some users who have sensitive data on their devices "expect" a format in recovery to be a full secure format.

I wouldn't disable something because "it takes too long" unless there was a real reason behind it.

The G2 may have a nand chip which is affected by some sort of secure erase bug.  I'm not sure why they would turn this flag on.

Patch-set: 1