Update patch set 1
Patch Set 1:

I'm just abusing the hooks that were put in for the Oppo N1 that allows a device to provide an extra settings option for controlling touchscreen gestures. It's not perfect, since the pick up gesture is tied to ambient display, but it's better than no config option for these at all.

Patch-set: 1