Update patch set 1
Patch Set 1:

extract-files.sh and setup-makefiles.sh in the device (ghost for me, but also at least obake) simply set the device and vendor variables and call the versions in msm8960dt-common.  The proprietary files defined in the device specific file never get pulled since the common versions of the scripts never look for them.  I would guess this happened when everything was split out in to the common project?  As stated, msm8225-common handles it properly and is what I based this modification on.

Patch-set: 1