Update patch set 3
Patch Set 3:

I checked for obake, which still did not get its LP update or a soak yet.

It is otherwise stated to arrive "soon", but not sure where to go next. We can either wait a little more, or I can look a bit into the firmware and see if ghost's can just be patched to run on obake (I'd expect only some data to differ).

Patch-set: 3