Update patch set 3
Patch Set 3:

Adding sys_boot allows the binary to reboot the phone. There's something wrong with the code that causes it to immediately fall into the case where the phone should be rebooted. If you remove sys_boot, the charger binary can't reboot the phone anymore, so it appears to 'work'.

Patch-set: 3